Changes to allow this to run.
Require(plyr) added earlier because revalue is needed within the corrgram block

In getTitle, the '+' was moved up a line to ensure that R recognized that the RHS of equation covered two lines.

'Noble' added to list of factors, so that when titles are consolidated later, 'Noble' will be a allowed target value.